move empty id validation to a shared function
