+++ internal/utils/debug.go
@@ -1,0 +1,11 @@
+package utils
+
+import (
+	"encoding/json"
+)
+
+// PancicPrettyPrint converts the interface into a json string and panics
+func PancicPrettyPrint(v interface{}) {
+	s, _ := json.MarshalIndent(v, "", "    ")
+	panic(string(s))
+}
