+++ internal/utils/utils.go
@@ -103,19 +103,3 @@ func ListIDsToModify(
 
 	return
 }
-
-// ListTypeAs converts a types.List into a known []T.
-func ListTypeAs[T any](
-	ctx context.Context,
-	list types.List,
-	diags *diag.Diagnostics,
-) []T {
-	if !IsKnown(list) {
-		return nil
-	}
-
-	var elements []T
-
-	diags.Append(list.ElementsAs(ctx, &elements, false)...)
-	return elements
-}
